Transform byte values into floats for caching

+++ src/main/java/com/gtnewhorizons/angelica/glsm/states/Color4.java
@@ -1,9 +1,9 @@
 package com.gtnewhorizons.angelica.glsm.states;
 
 public class Color4 {
-    public double red = 1.0F;
-    public double green = 1.0F;
-    public double blue = 1.0F;
-    public double alpha = 1.0F;
+    public float red = 1.0F;
+    public float green = 1.0F;
+    public float blue = 1.0F;
+    public float alpha = 1.0F;
 
 }
